// shared/types/village.d.ts
export type Building_Fa =
    | 'ساختمان اصلي'
    | 'دیوار'
    | 'انبار'
    | 'انبار عذا'
    | 'اردوگاه'
    | 'سفارتخانه'
    | 'عمارت قهرمان'
    | 'سربازخانه'
    | 'بازار'
    | 'دارالفنون'
    | 'قصر'
    | 'آهنگري'
    | 'اصطبل'
    | 'خزانه داري'
    | 'آجرپزي'
    | 'چوب بری'
    | 'ذوب آهن'
    | 'کارگاه'
    | 'آسياب'
    | 'ميدان تمرين'
    | 'نانوايي'
    | 'تالار شهر';

export type Building =
    | 'Main_Building'
    | 'Wall'
    | 'Warehouse'
    | 'Food_Storage'
    | 'Camp'
    | 'Embassy'
    | 'Hero_Mansion'
    | 'Barracks'
    | 'Market'
    | 'Academy'
    | 'Palace'
    | 'Blacksmith'
    | 'Stable'
    | 'Treasury'
    | 'Brick_Kiln'
    | 'Sawmill'
    | 'Iron_Foundry'
    | 'Workshop'
    | 'Mill'
    | 'Training_Ground'
    | 'Bakery'
    | 'Town_Hall';

export type PlaceId = 19 | 20 | 21 | 22 | 23 | 24 | 25 | 26 | 27 | 28 | 29 | 30 | 31 | 32 | 33 | 34 | 35 | 36 | 37 | 38 | 39 | 40;